Ensure installation directory is available
To avoid  the below error and to ensure that installation directory is
available , it's possible that user can specify the directory that
doesn't exists and expect the module to do it when needed. This change
takes care of it.

Could not retrieve catalog from remote server: Error 400 on SERVER:
Syntax error at '}'; expected ']' at
/etc/puppetlabs/puppet/modules/jdk_oracle/manifests/init.pp:103 on node