"Ensure installation directory is available""
To avoid  the below error and to ensure that installation directory is
available , it's possible that user can specify the directory that
doesn't exists and expect the module to do it when needed. This change
takes care of it.

change from notrun to 0 failed: Working directory '/usr/java' does not
exist